- Exception created and added to DAOs - Data access over Transactional Services that are pointing to daos - Bugs in DAOs fixed - DAOs save-method now returns the Bean with the generated ID

+++ tita-datasource/src/main/java/at/ac/tuwien/ifs/tita/datasource/dao/BaseDAO.java
@@ -6,6 +6,7 @@ import javax.persistence.PersistenceContext;
 import org.springframework.stereotype.Repository;
 
 import at.ac.tuwien.ifs.tita.datasource.domain.BaseEntity;
+import at.ac.tuwien.ifs.tita.datasource.exception.TitaDAOException;
 
 /**
  * 
@@ -37,11 +38,23 @@ public class BaseDAO<DomainClass extends BaseEntity> implements
      * @param entity
      *            - A sub-type of BaseEntity
      */
-    public void save(DomainClass entity) {
-        if (entity.getId() != null) {
-            entityManager.persist(entity);
-        } else {
-            entityManager.merge(entity);
+    public DomainClass save(DomainClass entity) throws TitaDAOException {
+        if (entity == null) {
+            throw new TitaDAOException(
+                    "Entity with value null cannot be persisted");
+        }
+
+        try {
+            if (entity.getId() == null) {
+                DomainClass retDomain = entityManager.merge(entity);
+                entityManager.persist(retDomain);
+                return retDomain;
+            } else {
+                return entityManager.merge(entity);
+            }
+        } catch (Exception e) {
+            throw new TitaDAOException("Exception was thrown: \n" + "Cause: "
+                    + e.getCause() + "\n" + "Error: " + e.getMessage());
         }
     }
 
@@ -51,8 +64,18 @@ public class BaseDAO<DomainClass extends BaseEntity> implements
      * @param entity
      *            - A sub-type of BaseEntity
      */
-    public void update(DomainClass entity) {
-        entityManager.merge(entity);
+    public void update(DomainClass entity) throws TitaDAOException {
+        if (entity == null || entity.getId() == null) {
+            throw new TitaDAOException(
+                    "Entity cannot be updated if its ID or itself is null");
+        }
+
+        try {
+            entityManager.merge(entity);
+        } catch (Exception e) {
+            throw new TitaDAOException("Exception was thrown: \n" + "Cause: "
+                    + e.getCause() + "\n" + "Error: " + e.getMessage());
+        }
     }
 
     /**
@@ -61,8 +84,18 @@ public class BaseDAO<DomainClass extends BaseEntity> implements
      * @param entity
      *            - A sub-type of BaseEntity
      */
-    public void delete(DomainClass entity) {
-        entityManager.remove(entityManager.merge(entity));
+    public void delete(DomainClass entity) throws TitaDAOException {
+        if (entity == null || entity.getId() == null) {
+            throw new TitaDAOException(
+                    "Entity cannot be deleted if its ID or itself is null");
+        }
+
+        try {
+            entityManager.remove(entityManager.merge(entity));
+        } catch (Exception e) {
+            throw new TitaDAOException("Exception was thrown: \n" + "Cause: "
+                    + e.getCause() + "\n" + "Error: " + e.getMessage());
+        }
     }
 
     /**
@@ -73,8 +106,25 @@ public class BaseDAO<DomainClass extends BaseEntity> implements
      * @return the object
      */
     public DomainClass getById(Class<? extends DomainClass> persistentClass,
-            Long id) {
+            Long id) throws TitaDAOException {
 
-        return entityManager.find(persistentClass, id);
+        if (id == null) {
+            throw new TitaDAOException("Entity with id: null cannot be found");
+        }
+
+        DomainClass entity = null;
+        try {
+            entity = entityManager.find(persistentClass, id);
+        } catch (Exception e) {
+            throw new TitaDAOException("Exception was thrown: \n" + "Cause: "
+                    + e.getCause() + "\n" + "Error: " + e.getMessage());
+        }
+
+        if (entity == null) {
+            throw new TitaDAOException("Entity with id: " + id
+                    + " cannot be found");
+        }
+
+        return entity;
     }
 }
